refactor(frontend): refactor components
