+++ frontend/src/components/home/readingSessions/index.ts
@@ -1,1 +1,2 @@
 export * from "./ReadingSessionList.tsx"
+export * from "./Calendar.tsx"
